test(alert): migrate a11y
